Refactor : 투두 정보 조회 Dto 리팩토링
